Added combination filter helper
Added a new function to retrieve the list of titles given a combination of languages
that all must be present.
Current impl. uses a raw SQL syntax due to lack of understanding of how to do it with
the ORM. We can improve that later.

Function is generic so it can be used both on titles and books and for both languages and tags